everything is fixed and good now
