+++ Assets/Prefabs/Frogger/PLavaFloor.prefab
@@ -538,8 +538,8 @@ GameObject:
   - component: {fileID: 4526130616469755264}
   - component: {fileID: 247377896309572702}
   - component: {fileID: 4643834184259073818}
-  - component: {fileID: 5447542439306332082}
   - component: {fileID: 238855848225607718}
+  - component: {fileID: 6721202763382581854}
   m_Layer: 0
   m_Name: GameLava
   m_TagString: Untagged
@@ -644,19 +644,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: {x: 0.1, y: 0}
---- !u!114 &5447542439306332082
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7423152499294277337}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4412ec93e56075a4494a0b989c3f460c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  loseText: {fileID: 0}
 --- !u!65 &238855848225607718
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -670,6 +657,18 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 0.25195086, z: 1}
   m_Center: {x: 0, y: -0.37402457, z: 0}
+--- !u!114 &6721202763382581854
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7423152499294277337}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 26f5cb659b9f049468dae91a8a2b0e84, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &9192938717895455946
 GameObject:
   m_ObjectHideFlags: 0
