+++ Assets/Prefabs/Frogger/RLLavaFloor.prefab
@@ -14,7 +14,7 @@ GameObject:
   - component: {fileID: 442676526235277861}
   - component: {fileID: 6457381472232552390}
   - component: {fileID: 3851070459758838121}
-  - component: {fileID: 708321094355104325}
+  - component: {fileID: 8593151983530008034}
   m_Layer: 0
   m_Name: GameLava
   m_TagString: Untagged
@@ -125,7 +125,7 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 0.29017258, z: 1}
   m_Center: {x: 0, y: -0.3549137, z: 0}
---- !u!114 &708321094355104325
+--- !u!114 &8593151983530008034
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -134,10 +134,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 2658627490960403474}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4412ec93e56075a4494a0b989c3f460c, type: 3}
+  m_Script: {fileID: 11500000, guid: 26f5cb659b9f049468dae91a8a2b0e84, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  loseText: {fileID: 0}
 --- !u!1 &4799667041812761064
 GameObject:
   m_ObjectHideFlags: 0
