+++ Assets/Scenes/Frogger.unity
@@ -132,7 +132,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 2913367432902255216, guid: 9c636108beddc1840bb1916edd79c68a, type: 3}
       propertyPath: m_RootOrder
-      value: 2
+      value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 2913367432902255216, guid: 9c636108beddc1840bb1916edd79c68a, type: 3}
       propertyPath: m_LocalScale.x
@@ -960,7 +960,7 @@ RectTransform:
   - {fileID: 970907955}
   - {fileID: 1592182266}
   m_Father: {fileID: 0}
-  m_RootOrder: 9
+  m_RootOrder: 8
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -1031,7 +1031,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 10
+  m_RootOrder: 9
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!4 &195029092 stripped
 Transform:
@@ -1100,7 +1100,7 @@ Transform:
   - {fileID: 1747113834}
   - {fileID: 1744736523}
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!4 &256360922 stripped
 Transform:
@@ -1238,7 +1238,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 7
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 180, z: 0}
 --- !u!4 &333103724 stripped
 Transform:
@@ -2017,7 +2017,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 8
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &789345080
 MonoBehaviour:
@@ -2160,7 +2160,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 3
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 54.04, y: 0, z: 0}
 --- !u!4 &863992767 stripped
 Transform:
@@ -3332,7 +3332,7 @@ Transform:
   - {fileID: 491925194}
   - {fileID: 809149358}
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1053216565
 PrefabInstance:
@@ -4217,7 +4217,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!4 &1379684049 stripped
 Transform:
@@ -6402,63 +6402,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 2859194311967815122, guid: 0d15efd8964ef8049a74c0e319c631de, type: 3}
   m_PrefabInstance: {fileID: 1926469854}
   m_PrefabAsset: {fileID: 0}
---- !u!1001 &2106362827
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_RootOrder
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -0.01635557
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: -14.51944
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246072, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3305141677869246073, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
-      propertyPath: m_Name
-      value: FroggerPlayer Variant
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: cd1a39e99e75f6c419659004c2865388, type: 3}
 --- !u!4 &2106861277 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 6222855548651565177, guid: 026a6ec7e376aef43a2840035b8ff3b5, type: 3}
