--- Assets/Scripts/Frogger/LavaKill.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4412ec93e56075a4494a0b989c3f460c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
