+++ Assets/Scripts/UI/Lobby.cs
@@ -18,7 +18,7 @@ namespace UI
         public void StartGame()
         {
             inputManager.enabled = false;
-            SceneTransition.LoadScene(3);
+            SceneTransition.LoadScene(6);
         }
 
         public void Back()
